Allow users to search for names of whiskey in table
