+++ client/src/components/WhiskeyTable.jsx
@@ -1,98 +1,10 @@
-import { useState, useEffect } from 'react'
-
-import { useQuery, useLazyQuery } from '@apollo/client'
-import { GET_WHISKEYS } from '../utils/queries'
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCaretDown, faCaretUp } from '@fortawesome/free-solid-svg-icons'
 
-import LoadingSpinner from './LoadingSpinner'
-import Search from './SearchWhiskey'
-import Pagination from './Pagination'
 import WhiskeyItem from './WhiskeyItem'
 
-const WhiskeyTable = () => {
-  const [searchTerm, setSearchTerm] = useState('')
-  const [perPage, setPerPage] = useState(20)
-  const [currentPage, setCurrentPage] = useState(1)
-  const [sortByName, setSortByName] = useState('asc')
-  const [sortByScore, setSortByScore] = useState(null)
-
-  const { loading, error, data, fetchMore, refetch } = useQuery(GET_WHISKEYS, {
-    variables: { search: searchTerm, page: currentPage, perPage, sortByName, sortByScore },
-  })
-
-  const totalPages = Math.ceil(data?.whiskeys?.count / perPage) || 0
-
-  useEffect(() => {
-    fetchMore({
-      variables: {
-        page: currentPage,
-        perPage: perPage,
-      },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!fetchMoreResult) return prev
-        return { ...prev, whiskeys: fetchMoreResult.whiskeys }
-      },
-    })
-  }, [currentPage, perPage])
-
-  useEffect(() => {
-    refetch({ search: searchTerm, page: currentPage, perPage, sortByName, sortByScore })
-  }, [searchTerm, sortByName, sortByScore])
-
-  useEffect(() => {
-    setCurrentPage(1)
-  }, [searchTerm])
-
-  const handlePageChange = (newPage) => {
-    setCurrentPage(newPage)
-  }
-
-  const handlePerPageChange = (e) => {
-    const newPerPage = parseInt(e.target.value)
-    setPerPage(newPerPage)
-    setCurrentPage(1)
-  }
-
-  const handleSortByName = () => {
-    const newSortByName = sortByName === 'asc' ? 'desc' : 'asc'
-    setSortByName(newSortByName)
-    setSortByScore(null)
-    setCurrentPage(1)
-  }
-
-  const handleSortByScore = () => {
-    const newSortByScore = sortByScore === 'highToLow' ? 'lowToHigh' : 'highToLow'
-    setSortByScore(newSortByScore)
-    setSortByName(null)
-    setCurrentPage(1)
-  }
-
-  const handleSearch = (value) => {
-    setSearchTerm(value)
-  }
-
-  if (loading) return <LoadingSpinner />
-
+const WhiskeyTable = ({ data, sortByName, sortByScore, handleSortByName, handleSortByScore, perPage }) => {
   return (
-    <section className="whiskeys">
-      <div className="flex justify-between mb-4">
-        <Search onSearch={handleSearch} />
-        <label htmlFor="perPage" className="mr-2">
-          Whiskeys per page:
-        </label>
-        <select
-          id="perPage"
-          className="p-1 border border-gray-300 rounded"
-          value={perPage}
-          onChange={handlePerPageChange}
-        >
-          <option value={20}>20</option>
-          <option value={50}>50</option>
-          <option value={100}>100</option>
-        </select>
-      </div>
       <table className="min-w-full divide-y divide-gray-200">
         <thead>
           <tr>
@@ -132,8 +44,6 @@ const WhiskeyTable = () => {
           ))}
         </tbody>
       </table>
-      <Pagination currentPage={currentPage} totalPages={totalPages} onPageChange={handlePageChange} />
-    </section>
   )
 }
